typescript setup and prototyping
