+++ package.json
@@ -31,10 +31,13 @@
     "xml-loader": "^1.2.1"
   },
   "dependencies": {
+    "@types/react": "^16.9.19",
+    "@types/three": "^0.103.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-three-fiber": "^4.0.8",
     "three": "^0.112.1",
-    "ts-loader": "^6.2.1"
+    "ts-loader": "^6.2.1",
+    "typescript": "^3.7.5"
   }
 }
