+++ tsconfig.json
@@ -11,7 +11,7 @@
     "module": "esnext", // Specify module code generation
     "moduleResolution": "node", // Resolve modules using Node.js style
     "resolveJsonModule": true, // Include modules imported with .json extension
-    "noEmit": true, // Do not emit output (meaning do not compile code, only perform type checking)
+    // "noEmit": true, // Do not emit output (meaning do not compile code, only perform type checking)
     "jsx": "react", // Support JSX in .tsx files
     "sourceMap": true, // *** Generate corrresponding .map file ***
     "declaration": true, // *** Generate corresponding .d.ts file ***
